requirements.txt
tensorflow>=2.10.0
keras>=2.10.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0
matplotlib>=3.5.0